project: Add method to get the longest hamiltonian path
Essentially the best we can do here is backtracking, so this is
still O(2^N). However, this is a somewhat interesting take on
the problem by using a fibres-like architecture, where we should
find the longest hamiltonian path a lot faster. However, right now
we are still exploring all the other paths "just in case".